@@ -7,7 +7,7 @@ apply plugin: 'idea'
 apply plugin: 'java'
 apply plugin: 'edu.sc.seis.launch4j'
 
-sourceCompatibility = 1.8
+sourceCompatibility = 14
 
 def mainClassPath = 'io.github.vampirestudios.tdg.start.TutLauncher'
 def exeJarName = rootProject.name + "-" + version
@@ -54,6 +54,8 @@ repositories {
     maven { url "http://repo.fusesource.com/nexus/content/repositories/releases-3rd-party/" }
     maven { url "https://oss.sonatype.org/content/repositories/snapshots/" }
     maven { url "https://oss.sonatype.org/content/repositories/releases/" }
+    maven { url "https://raw.github.com/SpinyOwl/repo/releases" }
+    maven { url "https://raw.github.com/SpinyOwl/repo/snapshots" }
 }
 
 launch4j {
@@ -77,35 +79,19 @@ dependencies {
     implementation "org.apache.logging.log4j:log4j-core:2.8.1"
 
     compile 'com.github.CoffeeCatRailway:CoffeeCatUtils:master-SNAPSHOT'
-    compile "net.fabricmc:fabric-loader:0.4.8+build.154"
     compile group: 'com.google.code.gson', name: 'gson', version: '2.8.5'
     compile group: 'com.google.common', name: 'google-collect', version: '1.0'
     compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.9'
 
-//    compile "org.lwjgl:lwjgl:$lwjglVersion"
-//    compile "org.lwjgl:lwjgl-glfw:$lwjglVersion"
-//    compile "org.lwjgl:lwjgl:$lwjglVersion:natives-windows"
-//    compile "org.lwjgl:lwjgl-glfw:$lwjglVersion:natives-windows"
-//    compile "org.lwjgl:lwjgl:$lwjglVersion:natives-linux"
-//    compile "org.lwjgl:lwjgl-glfw:$lwjglVersion:natives-linux"
-//    compile "org.lwjgl:lwjgl:$lwjglVersion:natives-macos"
-//    compile "org.lwjgl:lwjgl-glfw:$lwjglVersion:natives-macos"
-    compile "com.badlogicgames.gdx:gdx:1.9.10"
-    compile "com.badlogicgames.gdx:gdx-tools:1.9.10"
+//    compile "org.liquidengine:legui:2.2.0-SNAPSHOT"
+//    compile "org.liquidengine:leutil:1.0.0"
+//    compile "org.liquidengine:cbchain:1.0.0"
 
     // Netty
     compile group: 'io.netty', name: 'netty-all', version: '4.1.29.Final'
 
     testCompile group: 'junit', name: 'junit', version: '4.12'
     implementation "org.joml:joml:${jomlVersion}"
-}
-
-jar {
-    from configurations.compile.collect { zipTree it }
-    manifest {
-        attributes 'Main-Class': mainClassPath
-    }
-    exclude("META-INF/*.RSA", "META-INF/*.SF")
 }
 
 task copyOptions(type: Copy) {
