@@ -1,17 +1,18 @@
 package io.github.vampirestudios.tdg.objs;
 
 import coffeecatteam.coffeecatutils.NumberUtils;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonObject;
 import io.github.vampirestudios.tdg.gfx.Animation;
 import io.github.vampirestudios.tdg.gfx.assets.Assets;
 import io.github.vampirestudios.tdg.gfx.image.SpriteSheet;
 import io.github.vampirestudios.tdg.objs.entities.Entity;
 import io.github.vampirestudios.tdg.objs.tiles.Tiles;
 import io.github.vampirestudios.tdg.utils.Identifier;
-import org.json.simple.JSONArray;
-import org.json.simple.JSONObject;
 import org.newdawn.slick.Image;
 
 import java.util.HashMap;
+import java.util.Map;
 
 /**
  * @author CoffeeCatRailway
@@ -19,77 +20,70 @@ import java.util.HashMap;
  */
 public class EntityDataParser extends DataParser<Entity> {
 
-    private HashMap<String, HashMap<String, Animation>> LOADED = new HashMap<>();
+    private Map<String, HashMap<String, Animation>> LOADED = new HashMap<>();
 
     public EntityDataParser() {
         super("entities");
     }
 
     @Override
-    Entity customLoadData(JSONObject data, Entity obj) {
+    public Entity customLoadData(JsonObject data, Entity obj) {
         Entity entity = obj.newCopy();
 
-        if (!LOADED.keySet().contains(entity.getId())) {
+        if (!LOADED.containsKey(entity.getId())) {
             DataTypes.Entity type = DataTypes.Entity.getByName(String.valueOf(data.get("type")));
             SpriteSheet texture = getTexture(data);
             logger.info("Loading entity of type [" + type + "-" + type.typeName + "] with id [" + entity.getName() + "]");
 
             HashMap<String, Animation> textures = new HashMap<>();
             switch (type) {
-                case STATIC:
-                    JSONObject sprites = (JSONObject) data.get("sprites");
+                case STATIC -> {
+                    if (data.has("sprites")) {
+                        JsonObject sprites = (JsonObject) data.get("sprites");
+                        for (Object key : sprites.keySet()) {
+                            JsonObject sprite = (JsonObject) sprites.get((String) key);
+                            SpriteSheet spriteTexture;
+                            if (sprite.has("texture")) spriteTexture = getTexture(sprite).copy();
+                            else spriteTexture = texture.copy();
 
-                    for (Object key : sprites.keySet()) {
-                        JSONObject sprite = (JSONObject) sprites.get(key);
-                        SpriteSheet spriteTexture;
-                        if (sprite.containsKey("texture")) spriteTexture = getTexture(sprite).copy();
-                        else spriteTexture = texture.copy();
-
-                        if (sprite.containsKey("length"))
-                            textures.put((String) key, getAnimation(spriteTexture.getSheet(), sprite));
-                        else
-                            textures.put((String) key, getAnimation(spriteTexture.getSheet(), sprite, false, "spritePos"));
+                            if (sprite.has("length")) textures.put((String) key, getAnimation(spriteTexture.getSheet(), sprite));
+                            else textures.put((String) key, getAnimation(spriteTexture.getSheet(), sprite, false, "spritePos"));
+                        }
                     }
-                    break;
-
-                case LIVING:
-                    JSONObject animations = (JSONObject) data.get("animations");
-
-                    JSONArray size = (JSONArray) data.get("spriteSize");
+                }
+                case CREATURE -> {
+                    JsonObject animations = (JsonObject) data.get("animations");
+                    JsonArray size = (JsonArray) data.get("spriteSize");
                     int width = NumberUtils.parseInt(size.get(0));
                     int height = NumberUtils.parseInt(size.get(1));
-
                     for (Object key : animations.keySet()) {
-                        JSONObject anim = (JSONObject) animations.get(key);
+                        JsonObject anim = (JsonObject) animations.get((String) key);
                         textures.put(String.valueOf(key), getAnimation(texture.getSheet().copy(), anim, width, height));
                     }
-                    break;
-
-                case CUSTOM:
-                    textures.put(entity.getId(), new Animation(Tiles.AIR.getTexture()));
-                    break;
+                }
+                case CUSTOM -> textures.put(entity.getId(), new Animation(Tiles.AIR.getTexture()));
             }
 
             entity.setTextures(textures);
             LOADED.put(entity.getId(), textures);
         } else {
-            logger.info("Getting textures for entity with id [" + entity.getName() + "]");
+            logger.info("Getting textures for entity with an id of [" + entity.getName().replace("\"", "") + "]");
             entity.setTextures(LOADED.get(entity.getId()));
         }
 
         return entity.newCopy();
     }
 
-    private SpriteSheet getTexture(JSONObject data) {
+    private SpriteSheet getTexture(JsonObject data) {
         SpriteSheet texture = new SpriteSheet(Assets.MISSING_TEXTURE);
-        if (data.containsKey("texture")) {
+        if (data.has("texture")) {
             String texturePath = "textures/";
 
-            if (data.get("texture") instanceof JSONArray) {
-                JSONArray texturePaths = (JSONArray) data.get("texture");
-                texturePath += String.valueOf(texturePaths.get(NumberUtils.getRandomInt(0, texturePaths.size() - 1)));
+            if (data.get("texture") instanceof JsonArray) {
+                JsonArray texturePaths = (JsonArray) data.get("texture");
+                texturePath += String.valueOf(texturePaths.get(NumberUtils.getRandomInt(0, texturePaths.size() - 1))).replace("\"'", "");
             } else {
-                texturePath += data.get("texture");
+                texturePath += data.get("texture").getAsString().replace("\"'", "");
             }
 
             texture = new SpriteSheet(new Identifier("maewil", texturePath));
@@ -97,34 +91,34 @@ public class EntityDataParser extends DataParser<Entity> {
         return texture;
     }
 
-    private Animation getAnimation(Image texture, JSONObject data) {
-        JSONArray size = (JSONArray) data.get("spriteSize");
+    private Animation getAnimation(Image texture, JsonObject data) {
+        JsonArray size = (JsonArray) data.get("spriteSize");
         int width = NumberUtils.parseInt(size.get(0));
         int height = NumberUtils.parseInt(size.get(1));
         return getAnimation(texture, data, width, height);
     }
 
-    private Animation getAnimation(Image texture, JSONObject data, int width, int height) {
+    private Animation getAnimation(Image texture, JsonObject data, int width, int height) {
         return getAnimation(texture, data, width, height, true, "startPos");
     }
 
-    private Animation getAnimation(Image texture, JSONObject data, boolean xyExact, String posKey) {
-        JSONArray size = (JSONArray) data.get("spriteSize");
+    private Animation getAnimation(Image texture, JsonObject data, boolean xyExact, String posKey) {
+        JsonArray size = (JsonArray) data.get("spriteSize");
         int width = NumberUtils.parseInt(size.get(0));
         int height = NumberUtils.parseInt(size.get(1));
         return getAnimation(texture, data, width, height, xyExact, posKey);
     }
 
-    private Animation getAnimation(Image texture, JSONObject data, int width, int height, boolean xyExact, String posKey) {
+    private Animation getAnimation(Image texture, JsonObject data, int width, int height, boolean xyExact, String posKey) {
         int length = 1;
-        if (data.containsKey("length"))
+        if (data.has("length"))
             length = NumberUtils.parseInt(data.get("length"));
 
         int speed = Animation.DEFAULT_SPEED;
-        if (data.containsKey("speed"))
+        if (data.has("speed"))
             speed = NumberUtils.parseInt(data.get("speed"));
 
-        JSONArray pos = (JSONArray) data.get(posKey);
+        JsonArray pos = (JsonArray) data.get(posKey);
         int startPosX = NumberUtils.parseInt(pos.get(0));
         int startPosY = NumberUtils.parseInt(pos.get(1));
 
