@@ -6,7 +6,6 @@ import io.github.vampirestudios.tdg.start.MaewilEngine;
 import io.github.vampirestudios.tdg.start.MaewilLauncher;
 import io.github.vampirestudios.tdg.utils.Identifier;
 import io.github.vampirestudios.tdg.utils.registry.Registries;
-import org.json.simple.parser.ParseException;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -14,7 +13,7 @@ import java.util.List;
 
 public class Items {
 
-    public static final List<Item> UPDATABLE_TIEMS = new ArrayList<>();
+    public static final List<Item> UPDATABLE_ITEMS = new ArrayList<>();
 
     /*
      * General
@@ -43,7 +42,7 @@ public class Items {
     public static Item BREAD, APPLE;
     public static Item RAW_PORK, COOKED_PORK, RAW_STEAK, COOKED_STEAK;
 
-    public static void init(MaewilEngine maewilEngine) throws IOException, ParseException {
+    public static void init(MaewilEngine maewilEngine) throws IOException {
         ItemDataParser parser = new ItemDataParser();
 
         /*
