@@ -4,11 +4,9 @@ import coffeecatteam.coffeecatutils.position.Vector2D;
 import io.github.vampirestudios.tdg.gfx.Text;
 import io.github.vampirestudios.tdg.gfx.assets.Assets;
 import io.github.vampirestudios.tdg.gfx.ui.ClickListener;
-import io.github.vampirestudios.tdg.gfx.ui.WidgetTitleRender;
 import io.github.vampirestudios.tdg.gfx.ui.button.WidgetButton;
 import io.github.vampirestudios.tdg.gfx.ui.hyperlink.WidgetHyperlink;
 import io.github.vampirestudios.tdg.gfx.ui.hyperlink.WidgetHyperlinkCopyright;
-import io.github.vampirestudios.tdg.screen.options.CharacterCustomizationScreen;
 import io.github.vampirestudios.tdg.start.MaewilEngine;
 import io.github.vampirestudios.tdg.start.MaewilLauncher;
 import io.github.vampirestudios.tdg.utils.DiscordHandler;
@@ -25,12 +23,12 @@ public class TitleScreen extends Screen {
         super(maewilEngineIn);
 
         int yOff = 20;
-        uiManager.addObject(new WidgetTitleRender());
+//        uiManager.addObject(new WidgetTitleRender());
 
         float sizeP = 64;
         float xp = MaewilLauncher.WIDTH / 4f - sizeP / 2f;
         float yp = MaewilLauncher.HEIGHT / 2f - sizeP / 2f;
-        uiManager.addObject(btnPlayer = new WidgetButton(maewilEngineIn, new Vector2D(xp, yp), maewilEngine.getPlayer().getTextures().get("idle"), 4.0f, 0, new ClickListener() {
+        /*uiManager.addObject(btnPlayer = new WidgetButton(maewilEngineIn, new Vector2D(xp, yp), maewilEngine.getPlayer().getTextures().get("idle"), 4.0f, 0, new ClickListener() {
             @Override
             public void onClick() {
                 // TODO: Add entity texture data alt support
@@ -40,9 +38,9 @@ public class TitleScreen extends Screen {
             @Override
             public void update(GameContainer container, StateBasedGame game, int delta) {
             }
-        }));
-        btnPlayer.setRenderBtnBG(false);
-        btnPlayer.setRenderBtnHover(false);
+        }));*/
+//        btnPlayer.setRenderBtnBG(false);
+//        btnPlayer.setRenderBtnHover(false);
 
         uiManager.addObject(new WidgetButton(maewilEngineIn, true, MaewilLauncher.HEIGHT / 2 - yOff, "Select Game", new ClickListener() {
             @Override
